projects/python/requirements.txt
azure-identity==1.23.0
azure-storage-file-datalake==12.21.0
black==25.9.0
deltalake==0.25.5
duckdb==1.4.1
pandas==2.3.3
plotly>=5.0.0
streamlit==1.50.0
tabulate==0.9.0